Require authentication for web API calls
Two ways of authentication get added:

1. Password:
A password has to be set with the -s flag.
One can access every function using this password.

2. id_token:
Token received from logging in over twitch with the following url (replace <client-id>).
https://api.twitch.tv/kraken/oauth2/authorize?response_type=id_token%20code&client_id=<client-id>&redirect_uri=http://localhost&scope=openid
The token will be in the redirected url, which will look like this:
http://localhost/?code=xxx&id_token=xxx&scope=openid